refactored id to be long, created compund views ShortTravelDetails and TravelListRow

+++ shared/src/main/java/com/bnet/shared/model/backend/Providable.java
@@ -1,6 +1,6 @@
 package com.bnet.shared.model.backend;
 
 public interface Providable {
-    int getId();
-    void setId(int id);
+    long getId();
+    void setId(long id);
 }
