+++ shared/src/main/java/com/bnet/shared/model/backend/ProvidableRepository.java
@@ -3,7 +3,7 @@ package com.bnet.shared.model.backend;
 import java.util.List;
 
 public interface ProvidableRepository<T extends Providable> {
-    int addAndReturnAssignedId(T item);
+    long addAndReturnAssignedId(T item);
     List<T> getAll();
     List<T> getAllNews();
     T getOrNull(long id);
