+++ shared/src/main/java/com/bnet/shared/model/entities/Business.java
@@ -3,7 +3,7 @@ package com.bnet.shared.model.entities;
 import com.bnet.shared.model.backend.Providable;
 
 public class Business implements Providable {
-    private int id = -1;
+    private long id = -1;
     private String name = "";
     private Address address = new Address();
     private String phone = "";
@@ -22,12 +22,12 @@ public class Business implements Providable {
     }
 
     @Override
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     @Override
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
